duplicate extractor identifier test

+++ src/main/kotlin/Fragment.kt
@@ -4,6 +4,6 @@ import com.fasterxml.jackson.annotation.JsonRootName
 data class Fragment(
     var file : String? = null,
     var line : Long?   = null,
-    var start: Long?   = null,
-    var end  : Long?   = null
+    var start: Int?   = null,
+    var end  : Int?   = null
 )
